modifies imports

+++ qaequilibrae/modules/public_transport_procedures/gtfs_importer.py
@@ -1,6 +1,5 @@
 from os.path import dirname, join, isfile
 from aequilibrae.project.database_connection import database_connection
-from aequilibrae.utils.db_utils import commit_and_close
 
 from qgis.PyQt import uic
 from qgis.PyQt.QtCore import Qt
@@ -26,7 +25,9 @@ class GTFSImporter(QDialog, FORM_CLASS):
         self.feeds = []
         self.done = 1
 
-        if isfile(join(qgis_project.project.project_base_path, "public_transport.sqlite")):
+        self.is_pt_database = isfile(join(qgis_project.project.project_base_path, "public_transport.sqlite"))
+
+        if self.is_pt_database:
             self.rdo_clear.setText(self.tr("Overwrite Routes"))
             self.rdo_keep.setText(self.tr("Add to Existing Routes"))
         else:
@@ -34,6 +35,7 @@ class GTFSImporter(QDialog, FORM_CLASS):
             self.rdo_clear.setText(self.tr("Create new route system"))
             self.rdo_keep.setVisible(False)
             self.rdo_clear.setChecked(True)
+
         self.setFixedHeight(380)
         self.items = [self.config_box, self.progress_box]
 
@@ -67,24 +69,23 @@ class GTFSImporter(QDialog, FORM_CLASS):
             item.setEnabled(not item.isEnabled())
         self.setFixedHeight(176)
 
-        if self.rdo_clear.isChecked():
-            if isfile(join(self.qgis_project.project.project_base_path, "public_transport.sqlite")):
-                self.pt_conn = database_connection("transit")
-                for table in [
-                    "agencies",
-                    "fare_attributes",
-                    "fare_rules",
-                    "fare_zones",
-                    "pattern_mapping",
-                    "route_links",
-                    "routes",
-                    "stop_connectors",
-                    "stops",
-                    "trips",
-                    "trips_schedule",
-                ]:
-                    self.pt_conn.execute(f"DELETE FROM {table};")
-                self.pt_conn.commit()
+        if self.rdo_clear.isChecked() and self.is_pt_database:
+            self.pt_conn = database_connection("transit")
+            for table in [
+                "agencies",
+                "fare_attributes",
+                "fare_rules",
+                "fare_zones",
+                "pattern_mapping",
+                "route_links",
+                "routes",
+                "stop_connectors",
+                "stops",
+                "trips",
+                "trips_schedule",
+            ]:
+                self.pt_conn.execute(f"DELETE FROM {table};")
+            self.pt_conn.commit()
 
         for _, feed in enumerate(self.feeds):
             feed.signal.connect(self.signal_handler)
@@ -92,14 +93,12 @@ class GTFSImporter(QDialog, FORM_CLASS):
             feed.set_allow_map_match(True)
             feed.doWork()
 
-            with commit_and_close(database_connection("transit")) as conn:
-                list = [col for col in conn.execute("PRAGMA table_info(routes)").fetchall()]
-                print(list)
-                if "v_route_id" not in list:
-                    conn.execute("alter TABLE routes add v_route_id text")
-                conn.execute("update routes set v_route_id=cast(route_id as text) where v_route_id is null")
-
         self.close()
+
+        for layer in ["transit_routes", "transit_stops"]:
+            self.qgis_project.create_layer_by_name(layer)
+
+        print(self.qgis_project.layers.keys())
 
     def signal_handler(self, val):
         if len(val) == 1:
